LayoutTests/canvas/philip/tests/2d.text.font.parse.basic.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.font.parse.basic</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

ctx.font = '20px serif';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");

ctx.font = '20PX   SERIF';
_assertSame(ctx.font, '20px serif', "ctx.font", "'20px serif'");


});
</script>
